+++ src/components/Home/Header.astro
@@ -1,5 +1,5 @@
 ---
-import { Search, SearchMobile } from './Search'
+import { Search, SearchMobile } from '@/components/Search'
 import { Search as SearchIcon, Telescope as PopcornIcon } from 'lucide-astro'
 
 const pathname = Astro.url.pathname
