+++ src/components/Media/Discover.astro
@@ -1,5 +1,5 @@
 ---
-import DiscoverMovie from '@/components/DiscoverCard.astro'
+import DiscoverMovie from '@/components/Media/DiscoverCard.astro'
 import { ChevronLeft, ChevronRight } from 'lucide-astro'
 import type { Movie } from 'tmdb-ts'
