+++ src/components/Media/DiscoverCard.astro
