+++ src/layouts/Layout.astro
@@ -1,6 +1,6 @@
 ---
-import Footer from '@/components/Footer.astro'
-import Header from '@/components/Header.astro'
+import Footer from '@/components/Home/Footer.astro'
+import Header from '@/components/Home/Header.astro'
 import { ViewTransitions } from 'astro:transitions'
 import '@/styles/globals.css'
