+++ src/pages/explore/index.astro
@@ -1,7 +1,7 @@
 ---
 import { Image } from 'astro:assets'
 import Layout from '@/layouts/Layout.astro'
-import { ExplorePage } from '@/components/Explore'
+import { ExplorePage } from '@/components/Explore/Explore'
 import { Search as SearchIcon } from 'lucide-astro'
 import { getTrending, multiSearch } from '@/utils/tmdb'
 import { getImagePath } from '@/utils'
