+++ src/pages/play/[media]/[mediaId]/[mediaTitle]/index.astro
@@ -1,8 +1,8 @@
 ---
-import MediaGrid from '@/components/MediaGrid.astro'
-import PlayMediaCard from '@/components/PlayMediaCard.astro'
-import SelectSource from '@/components/SelectSource.astro'
-import TVShowPlayer from '@/components/TVShowPlayer.astro'
+import MediaGrid from '@/components/Media/MediaGrid.astro'
+import PlayMediaCard from '@/components/Media/PlayMediaCard.astro'
+import SelectSource from '@/components/Player/SelectSource.astro'
+import TVShowPlayer from '@/components/Player/TVShowPlayer.astro'
 import Layout from '@/layouts/Layout.astro'
 import { getImagePath, getSeasonOrEpisode } from '@/utils'
 import { getMovieUrl, getSource } from '@/utils/sources'
