+++ src/utils/sources.ts
@@ -1,4 +1,4 @@
-interface Source {
+export interface Source {
   id: string
   name: string
   movieUrl: string
